refactor: centralize lora network logic
